feat: Updated image references for Redis and Alpine containers
- Motivation and Context

The image references for Redis containers in the east, north, and west custom
resources needed to be updated to use mirrored versions hosted on GCR for
improved reliability and scalability.

- How Has This Been Tested?

The changes were tested locally by deploying the updated CRs and verifying
that the containers were pulled from the correct mirrored images.

Signed-off-by: Rafael Zago <rafaelvzago@gmail.com>